[Workflow] Extract abstraction for extending workflow compiler passes
